+++ lib/h-timber.php
@@ -4,8 +4,18 @@ class H_Twig {
 
   function __construct() {
     add_filter('get_twig', array($this, 'add_to_twig') );
+
+    // enable password-protected post
+    add_filter('timber/post/content/show_password_form_for_protected', function($maybe_show) {
+      return true;
+    });
   }
 
+  /*
+    Add extra filter to Twig
+
+    @filter get_twig
+  */
   function add_to_twig($twig) {
     $twig->addExtension(new Twig_Extension_StringLoader() );
 
@@ -21,10 +31,18 @@ class H_Twig {
       return $date_ago;
     }) );
 
-    // Dump all methods available in the object
-    $twig->addFilter('methods', new Twig_Filter_Function(function($object) {
-      var_dump(get_class_methods($object) );
-    }) );
+    // only if DEBUG is true
+    if (defined('WP_DEBUG') && WP_DEBUG === true) {
+      // Dump the object
+      $twig->addFilter('dump', new Twig_Filter_Function(function($object) {
+        var_dump($object);
+      }) );
+
+      // Dump all methods available in the object
+      $twig->addFilter('methods', new Twig_Filter_Function(function($object) {
+        var_dump(get_class_methods($object) );
+      }) );
+    }
 
     return $twig;
   }